# src/main/resources/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Web VolunteerDTO Coordination System - OpenAPI 3.0
  description: |-
    This is a Web Volunteer Coordination System based on the OpenAPI 3.0 specification. 
  contact:
    email: 260292@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: localhost:8080/
...
paths:
  /event:
    post:
      tags:
        - event
      summary: Add a new event
      description: Add a new event
      operationId: addEvent
      requestBody:
        description: Create a new event
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDTO'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDTO'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
    get:
      tags:
        - event
      summary: Finds all events
      description: Lists all events
      operationId: findAllEvents
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /event/findInTimeframe:
    get:
      tags:
        - event
      summary: Finds events in given timeframe
      description: Lists all events between two passed dates
      operationId: findInTimeframe
      parameters:
        - name: before
          in: query
          description: Earliest inclusive date of event start
          schema:
            type: string
            format: date-time
        - name: after
          in: query
          description: Latest inclusive date of event start
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
  /event/findByEventTags:
    get:
      tags:
        - event
      summary: Finds events by event tags
      description: Multiple tags can be provided with comma separated strings.
      operationId: findEventsByEventTags
      parameters:
        - name: tagIds
          in: query
          description: Tags to filter by
          required: false
          explode: true
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
        '400':
          description: Invalid tag
        '500':
          description: Internal server error
  /event/{eventId}:

    put:
      tags:
        - event
      summary: Update an existing event
      description: Update an existing event by Id
      operationId: updateEvent
      parameters:
        - name: eventId
          in: path
          description: ID of event to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update an existent event
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventWithoutTagsDTO'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventWithoutTagsDTO'
            application/xml:
              schema:
                $ref: '#/components/schemas/EventWithoutTagsDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: event not found
        '422':
          description: Validation exception
    get:
      tags:
        - event
      summary: Get detailed event information
      description: Returns a single detailed event
      operationId: getEventById
      parameters:
        - name: eventId
          in: path
          description: ID of event to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDTO'
        '404':
          description: EventDTO not found
        '500':
          description: Internal server error
    delete:
      tags:
        - event
      summary: Deletes an event
      description: Deletes an event
      operationId: deleteEvent
      parameters:
        - name: eventId
          in: path
          description: event id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '404':
          description: EventDTO not found
        '500':
          description: Internal server error
#  /event/{eventId}/uploadImage:
#    post:
#      tags:
#        - event
#      summary: uploads an image
#      description: ''
#      operationId: uploadFile
#      parameters:
#        - name: eventId
#          in: path
#          description: ID of event to update
#          required: true
#          schema:
#            type: integer
#            format: int64
#      requestBody:
#        content:
#          image/png:
#            schema:
#              type: string
#              format: binary
#      responses:
#        '200':
#          description: successful operation
#        '400':
#          description: Invalid input
#        '404':
#          description: EventDTO not found
#        '500':
#          description: Internal server error
...
  /eventtag:
    get:
      tags:
        - eventtag
      summary: Retrieve all event tags
      operationId: getAllEventTags
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventTagDTO'

    post:
      tags:
        - eventtag
      summary: Create a new event tag
      operationId: createEventTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventTagDTO'
      responses:
        '201':
          description: EventDTO tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTagDTO'
        '400':
          description: EventDTO tag already exists
        '500':
          description: Internal server error
  /eventtag/{eventTagId}:
    delete:
      tags:
        - eventtag
      summary: Delete an event tag
      operationId: deleteEventTag
      parameters:
        - name: eventTagId
          in: path
          required: true
          description: ID of the event tag to delete
          schema:
            type: integer
      responses:
        '204':
          description: EventDTO tag deleted successfully
        '404':
          description: EventDTO tag not found
        '500':
          description: Internal server error
...
  /crew/{volunteerId}/{eventId}:
    post:
      tags:
        - crew
      summary: Join an event crew
      operationId: joinEventCrew
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of the volunteer joining the crew
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          description: ID of the event for which the volunteer is joining the crew
          schema:
            type: integer
      responses:
        '201':
          description: VolunteerDTO successfully joined the event crew
        '400':
          description: Bad request

    put:
      tags:
        - crew
      summary: Update crew status
      operationId: updateCrewStatus
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of volunteer to update their status
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          description: ID of event to update
          schema:
            type: integer
        - name: status
          in: query
          required: true
          description: New status of the crew
          schema:
            type: string
            enum: [busy, gone, available]
      responses:
        '200':
          description: CrewDTO status updated successfully
        '400':
          description: Invalid input
        '404':
          description: CrewDTO member not found
        '500':
          description: Internal server error
    get:
      tags:
        - crew
      summary: Generate a certificate
      operationId: getCrewCertificate
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of volunteer to generate certificate
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          description: ID of event to retrieve data from
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewWithCertificateDTO'
        '404':
          description: CrewDTO member not found

    delete:
      tags:
        - crew
      summary: Delete a crew member
      operationId: deleteCrew
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of volunteer to delete from crew
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          description: ID of event to remove from
          schema:
            type: integer
      responses:
        '204':
          description: CrewDTO deleted successfully
        '404':
          description: CrewDTO member not found
  /crew/event/{eventId}:
    get:
      tags:
        - crew
      summary: Get all crew members for a given eventId
      operationId: getAllCrewsByEventId
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event to retrieve crews for
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrewDTO'
        '404':
          description: CrewDTO members not found
        '500':
          description: Internal server error
  /crew/volunteer/{volunteerId}:
    get:
      tags:
        - crew
      summary: Get all crew memberships for a given volunteer
      operationId: getAllCrewsByVolunteerId
      parameters:
        - name: volunteerId
          in: path
          required: true
          description: ID of the volunteer to retrieve crews for
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrewDTO'
        '404':
          description: CrewDTO members not found
        '500':
          description: Internal server error
components:
  schemas:
    EventDTO:
      required:
        - name
        - date
        - organizerId
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Juwenalia 2024
        date:
          type: string
          format: date-time
          example: 2024-05-01T12:00:00Z
        organizerId:
          type: integer
          format: int64
          example: 10
        eventTags:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/EventTagDTO'
        mapPath:
          type: string
          example: Juwenalia_map.png
          description: event location map image path
      xml:
        name: event
    EventWithoutTagsDTO:
      required:
        - name
        - date
        - organizerId
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Juwenalia 2024
        date:
          type: string
          format: date-time
          example: 2024-05-01T12:00:00Z
        organizerId:
          type: integer
          format: int64
          example: 198772
        mapPath:
          type: string
          example: Juwenalia_map.png
          description: event location map image path
      xml:
        name: eventwithouttags
...
    VolunteerDTO:
      type: object
      required:
        - name
        - surname
        - username
        - userId
      properties:
        id:
          type: integer
          format: int64
          example: 102856
        name:
          type: string
          example: John
        surname:
          type: string
          example: Doe
        username:
          type: string
          example: joey
        userId:
          type: integer
          format: int64
          example: 100320
      xml:
        name: volunteer
...
    EventTagDTO:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Concert
      xml:
        name: eventtag
...
    CrewDTO:
      type: object
      properties:
        volunteer:
          $ref: '#/components/schemas/VolunteerDTO'
        eventId:
          type: integer
          format: int64
        status:
          type: string
          enum: [busy, gone, available]
      xml:
        name: crew
    CrewWithCertificateDTO:
      type: object
      properties:
        volunteer:
          $ref: '#/components/schemas/VolunteerDTO'
        eventId:
          type: integer
          format: int64
        certificatePath:
          type: string
          example: joey_Juwenalia 2024.pdf
      xml:
        name: crewWithCertificate
...